pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jspecter"
version = "1.0.0"
description = "Autonomous JavaScript Recon, Secret Discovery & Vulnerability Intelligence Engine"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "JSPECTER Contributors" }]
keywords = ["security", "bugbounty", "javascript", "recon", "cve", "pentest", "react2shell"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]
requires-python = ">=3.9"
dependencies = ["aiohttp>=3.9.0", "aiofiles>=23.0.0"]

[project.optional-dependencies]
git = ["gitpython>=3.1.0"]
dev = ["pytest>=7.0", "pytest-asyncio>=0.21", "black>=23.0", "mypy>=1.0", "flake8>=6.0"]

[project.scripts]
jspecter = "jspecter.cli:main"

[project.urls]
Homepage = "https://github.com/abhi04anon/jspecter"
"Bug Reports" = "https://github.com/abhi04anon/jspecter/issues"
Source = "https://github.com/abhi04anon/jspecter"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
